Fixes & Features
-Fixed Invincium model issues.
-Re-added Glow Glass. IgnoreSimilarity isn't working for some reason, so
that'll need fixing...
-Re-added Godstone. Now deals magic damage to undead mobs, meaning it
can kill Zombie Pigmen, Wither Skeletons, and Withers in addition to
normal Zombies and Skeletons (before it would light those mobs on fire,
but since they're immune to fire they wouldn't take any damage).
-Godstone now removes all harmful potion effects from any living player
or mob that stands near it.
-Mobs damaged by Godstone emit a particle effect relative to the damage
they recieve (the minimum being 0.5 hearts and the max being 4).
-Added method to BlockBase to specify a block's slipperiness level.
-Tried to add a method to specify whether the block can be destroyed by
the wither/ender dragon, but it's not working for whatever reason...
-Added Building Blocks creative tab.
-Added the "alternate vanilla textures" resource pack to the assets
folder. Ideally, we'd have the mod auto-download it upon installation
and set it as the selected resource pack by default, but hell if I know
how to do that, lol.