i want to gouge my eyes out with a spork
-Tried to fix the fucking multistate itemBlock bug.
-Tried to fix the fucking multistate itemBlock bug again.
-This process repeated for three fucking hours.
-I’ve successfully managed to get to a point where two Cold Sand blocks
show up in the creative inventory, each with their own metadata value.
However, both of them have the same texture, and both of them place the
default state of the Cold Sand block.
-You can summon in Cold Red Sand just fine using /setblock ~ ~ ~
minestrapp:cold_sand 1. The state clearly exists, so at least I know
for sure it’s not a problem with the block states not registering.